Исправления Вынесение ошибок в отдельный файл Изменение иерархии директорий
