+++ pkg/leetcode/two-sum/errors.go
@@ -1,0 +1,7 @@
+package twosum
+
+import "errors"
+
+var (
+	ErrInvalidNumbers = errors.New("invalid numbers")
+)
